requirements/dev.txt
-r base.txt
django-extensions==1.3.3
epc==0.0.5
ipdb==0.8
ipython==1.2.1
jedi==0.7.0
Werkzeug==0.9.4
pip-tools==0.3.4
django-debug-toolbar==1.3.2
